+++ weekly/PRB_StrainJAX2.py
@@ -113,10 +113,8 @@ def mmrrc():
     
     # Retrieve all Strains that have a MMRRC ID and whose Alleles are used in a Genotype
 
-    try:
+    if os.environ['DB_TYPE'] == 'postgres':
     	db.sql('drop table #strains', None)
-    except:
-	pass
 
     db.sql('''
 	select distinct s._Strain_key,
